게시글 수정, 삭제 완료

+++ web/view.jsp
@@ -117,7 +117,7 @@
             if (userId != null && userId.equals(board.getUserId())) {
         %>
         <a href="update.jsp?boardId=<%=boardId%>" class="btn btn-primary">수정</a>
-        <a href="deleteAction.jsp?boardId=<%=boardId%>" class="btn btn-primary">삭제</a>
+        <a onclick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?boardId=<%=boardId%>" class="btn btn-primary">삭제</a>
                 <%
         }
         %>
